automation
My code is an automation  as it automatically waits until a specified time (scheduled_time) to start playing a video, displays a "Video Finished!" message when the video playback is complete and automatically checks the laptop's battery usage and sends a notification when the usage exceeds a certain threshold (max_usage_duration). 
Its an automation as I have set the time(scheduled_time ) to 10 pm which will make the video to pop at the time so that I will be reminded of the time or  to take rest. Along  with the video I have also set up an automation where it will also be  reminding me to take rest if  I have been using the laptop for more than 1 hour after the video is played so basically at 11 pm.